add firestore function

+++ firestore/modules/chatting/firestore.ts
@@ -1,9 +1,9 @@
-
+// useLibs
 // noPage
 
 import _global from 'esoftplay/_global';
 import { initializeApp } from 'firebase/app';
-import { addDoc, collection, deleteDoc, doc, FieldPath, getDoc, getDocs, initializeFirestore, onSnapshot, query, setDoc, updateDoc, where, WhereFilterOp } from 'firebase/firestore';
+import { addDoc, collection, deleteDoc, doc, FieldPath, getDoc, getDocs, initializeFirestore, limit, onSnapshot, orderBy, OrderByDirection, query, setDoc, startAfter, updateDoc, where, WhereFilterOp } from 'firebase/firestore';
 
 export interface DataId {
   id: string,
@@ -11,6 +11,7 @@ export interface DataId {
 }
 
 export type id = string
+let lastVisible: any = null
 
 const Firestore = {
   init() {
@@ -22,7 +23,7 @@ const Firestore = {
       storageBucket: "esoftplay-log.appspot.com",
       messagingSenderId: "844016531377",
       appId: "1:844016531377:web:892688387299a7b0d3af90"
-    })
+    }, "firestore")
     // _global.firebaseApp = initializeApp({
     //   "apiKey": "AIzaSyB04JT4JJfFsArIccAjBEn1nwIlg8EVWx4",
     //   "authDomain": "bigbang-online.firebaseapp.com",
@@ -42,7 +43,7 @@ const Firestore = {
   add: {
     doc(path: string[], value: any, cb: () => void, err?: (error: any) => void) {
       if (path.length % 2 > 0) {
-        console.warn("path untuk akses Doc data tidak boleh berhenti di Collection")
+        console.warn("path untuk akses Doc data tidak boleh berhenti di Collection [Firestore.add.doc]")
         return
       }
       const colRef = doc(Firestore.db(), ...path)
@@ -52,7 +53,7 @@ const Firestore = {
     },
     collection(path: string[], value: any, cb: (dt: any) => void, err?: (error: any) => void) {
       if (path.length % 2 == 0) {
-        console.warn("path untuk akses Collection data tidak boleh berhenti di Doc")
+        console.warn("path untuk akses Collection data tidak boleh berhenti di Doc [Firestore.add.collection]")
         return
       }
       //@ts-ignore
@@ -64,6 +65,10 @@ const Firestore = {
   },
   delete: {
     doc(path: string[], cb: () => void, err?: (error: any) => void) {
+      if (path.length % 2 > 0) {
+        console.warn("path untuk akses Doc data tidak boleh berhenti di Collection [Firestore.delete.doc]")
+        return
+      }
       const colRef = doc(Firestore.db(), ...path)
       deleteDoc(colRef).then((snap) => {
         cb()
@@ -73,26 +78,41 @@ const Firestore = {
   get: {
     doc(path: string[], condition: [fieldPath?: string | FieldPath, opStr?: WhereFilterOp, value?: unknown], cb: (arr: DataId) => void, err?: (error: any) => void) {
       if (path.length % 2 > 0) {
-        console.warn("path untuk akses Doc data tidak boleh berhenti di Collection")
+        console.warn("path untuk akses Doc data tidak boleh berhenti di Collection [Firestore.get.doc]")
         return
       }
       //@ts-ignore
       const colRef = doc(Firestore.db(), ...path)
       //@ts-ignore
       const fRef = (!condition || condition.length < 3) ? colRef : query(colRef, where(...condition))
+      //@ts-ignore
       getDoc(fRef).then((snap) => {
         cb({ data: snap.data(), id: snap.id })
       }).catch(err)
     },
-    collection(path: string[], condition: [fieldPath?: string | FieldPath, opStr?: WhereFilterOp, value?: unknown], cb: (arr: DataId[]) => void, err?: (error: any) => void) {
+    collectionWhereOrderBy(path: string[], condition: [fieldPath?: string | FieldPath, opStr?: WhereFilterOp, value?: unknown][], order_by: [fieldPath?: string | FieldPath, directionStr?: OrderByDirection | undefined][], cb: (arr: DataId[]) => void, err?: (error: any) => void) {
       if (path.length % 2 == 0) {
-        console.warn("path untuk akses Collection data tidak boleh berhenti di Doc")
+        console.warn("path untuk akses Collection data tidak boleh berhenti di Doc [Firestore.get.collection]")
         return
       }
       //@ts-ignore
       const colRef = collection(Firestore.db(), ...path)
-      //@ts-ignore
-      const fRef = (!condition || condition.length < 3) ? colRef : query(colRef, where(...condition))
+      let conditionsArray: any = []
+      condition.forEach((c) => {
+        //@ts-ignore
+        conditionsArray.push(where(...c))
+      })
+      let orderArray: any = []
+      order_by.forEach((o) => {
+        //@ts-ignore
+        orderArray.push(orderBy(...o))
+      })
+      // @ts-ignore
+      // const fRef = (!condition || condition.length == 0) ? colRef : query(colRef, ...conditionsArray)
+      const fRef = (condition && order_by) ? query(colRef, ...conditionsArray, ...orderArray)
+        : condition ? query(colRef, ...conditionsArray)
+          : order_by ? query(colRef, ...orderArray) : colRef
+
       let datas: any[] = []
       getDocs(fRef).then((snap) => {
         snap.docs.forEach((doc) => {
@@ -103,7 +123,7 @@ const Firestore = {
     },
     collectionIds(path: string[], condition: [fieldPath?: string | FieldPath, opStr?: WhereFilterOp, value?: unknown], cb: (arr: id[]) => void, err?: (error: any) => void) {
       if (path.length % 2 == 0) {
-        console.warn("path untuk akses Collection data tidak boleh berhenti di Doc")
+        console.warn("path untuk akses Collection data tidak boleh berhenti di Doc [Firestore.get.collectionIds]")
         return
       }
       //@ts-ignore
@@ -117,14 +137,42 @@ const Firestore = {
         })
         cb(datas)
       }).catch(err)
-    }
+    },
+    collection(path: string[], cb: (arr: DataId[]) => void, err?: (error: any) => void) {
+      Firestore.get.collectionWhereOrderBy(path, [], [], cb, err)
+    },
+    collectionWhere(path: string[], condition: [fieldPath?: string | FieldPath, opStr?: WhereFilterOp, value?: unknown][], cb: (arr: DataId[]) => void, err?: (error: any) => void) {
+      Firestore.get.collectionWhereOrderBy(path, condition, [], cb, err)
+    },
+    collectionOrderBy(path: string[], order_by: [fieldPath?: string | FieldPath, directionStr?: OrderByDirection | undefined][], cb: (arr: DataId[]) => void, err?: (error: any) => void) {
+      Firestore.get.collectionWhereOrderBy(path, [], order_by, cb, err)
+    },
   },
   listen: {
-    collection(path: string[], cb: (dt: any) => void, err?: (error: any) => void): () => void {
-      // @ts-ignore
+    collection(path: string[], condition: [fieldPath?: string | FieldPath, opStr?: WhereFilterOp, value?: unknown][], order_by: [fieldPath?: string | FieldPath, directionStr?: OrderByDirection | undefined][], cb: (dt: any) => void, err?: (error: any) => void): () => void {
+      if (path.length % 2 == 0) {
+        console.warn("path untuk akses Collection data tidak boleh berhenti di Doc [Firestore.listen.collection]")
+        return () => { }
+      }
+      //@ts-ignore
       const colRef = collection(Firestore.db(), ...path)
+      let conditionsArray: any = []
+      condition.forEach((c) => {
+        //@ts-ignore
+        conditionsArray.push(where(...c))
+      })
+      let orderArray: any = []
+      order_by.forEach((o) => {
+        //@ts-ignore
+        orderArray.push(orderBy(...o))
+      })
       let datas: any[] = []
-      const unsub = onSnapshot(colRef, (snap) => {
+
+      const fRef = (condition && order_by) ? query(colRef, ...conditionsArray, ...orderArray)
+        : condition ? query(colRef, ...conditionsArray)
+          : order_by ? query(colRef, ...orderArray) : colRef
+
+      const unsub = onSnapshot(fRef, (snap) => {
         datas = []
         snap.docs.forEach((doc) => {
           datas.push({ data: doc.data(), id: doc.id })
@@ -134,6 +182,10 @@ const Firestore = {
       return () => unsub()
     },
     doc(path: string[], cb: (dt: any) => void, err?: (error: any) => void): () => void {
+      if (path.length % 2 > 0) {
+        console.warn("path untuk akses Doc data tidak boleh berhenti di Collection [Firestore.listen.doc]")
+        return () => { }
+      }
       // @ts-ignore
       const colRef = doc(Firestore.db(), ...path)
       const unsub = onSnapshot(colRef, (snap) => {
@@ -143,16 +195,14 @@ const Firestore = {
     }
   },
   update: {
-    collection() {
-
-    },
-    doc(path: string[], value: string, cb: () => void, err?: (error: any) => void) {
-      let nPath = [path[0], path[1]]
-      let par = path.filter((t, i) => i != 0 && i != 1)
-
-      const colRef = doc(Firestore.db(), ...nPath)
+    doc(path: string[], index: string, value: string, cb: () => void, err?: (error: any) => void) {
+      if (path.length % 2 > 0) {
+        console.warn("path untuk akses Doc data tidak boleh berhenti di Collection [Firestore.update.doc]")
+        return
+      }
+      const colRef = doc(Firestore.db(), ...path)
       updateDoc(colRef, {
-        [`${par.join(".")}`]: value
+        [`${index}`]: value
       }).then((e) => {
         cb()
       }).catch(err)
@@ -161,9 +211,52 @@ const Firestore = {
   query() {
 
   },
-  paginate() {
+  paginate(isStartPage: boolean, path: string[], condition: [fieldPath?: string | FieldPath, opStr?: WhereFilterOp, value?: unknown][], order_by: [fieldPath?: string | FieldPath, directionStr?: OrderByDirection | undefined][], limitItem: number, cb: (dt: any, endReach: boolean) => void, err?: (error: any) => void): void {
+    if (path.length % 2 == 0) {
+      console.warn("path untuk akses Collection data tidak boleh berhenti di Doc [Firestore.paginate]")
+      return
+    }
+    //@ts-ignore
+    const colRef = collection(Firestore.db(), ...path)
 
-  }
+    let conditionsArray: any = []
+    condition.forEach((c) => {
+      //@ts-ignore
+      conditionsArray.push(where(...c))
+    })
+
+    let orderArray: any = []
+    order_by.forEach((o) => {
+      //@ts-ignore
+      orderArray.push(orderBy(...o))
+    })
+
+    const fRef = (conditionsArray.length > 0 && orderArray.length > 0) ? query(colRef, ...conditionsArray, ...orderArray, limit(limitItem))
+      : conditionsArray.length > 0 ? query(colRef, ...conditionsArray, limit(limitItem))
+        : orderArray.length > 0 ? query(colRef, ...orderArray, limit(limitItem)) : colRef
+
+    const fRef1 = (conditionsArray.length > 0 && orderArray.length > 0) ? query(colRef, ...conditionsArray, ...orderArray, startAfter(lastVisible || 0), limit(limitItem))
+      : conditionsArray.length > 0 ? query(colRef, ...conditionsArray, startAfter(lastVisible || 0), limit(limitItem))
+        : orderArray.length > 0 ? query(colRef, ...orderArray, startAfter(lastVisible || 0), limit(limitItem)) : colRef
+
+    let datas: any[] = []
+    getDocs(isStartPage ? fRef : fRef1).then((snap) => {
+      snap.docs.forEach((doc) => {
+        datas.push({ data: doc.data(), id: doc.id })
+      })
+      lastVisible = snap.docs[snap.docs.length - 1]
+      cb(datas, snap.empty)
+    }).catch(err)
+  },
+  paginateOrderBy(isStartPage: boolean, path: string[], order_by: [fieldPath?: string | FieldPath, directionStr?: OrderByDirection | undefined][], cb: (dt: any, endReach: boolean) => void, err?: (error: any) => void) {
+    Firestore.paginate(isStartPage, path, [], order_by, 20, cb, err)
+  },
+  paginateWhere(isStartPage: boolean, path: string[], condition: [fieldPath?: string | FieldPath, opStr?: WhereFilterOp, value?: unknown][], cb: (dt: any, endReach: boolean) => void, err?: (error: any) => void) {
+    Firestore.paginate(isStartPage, path, condition, [], 20, cb, err)
+  },
+  paginateLimit(isStartPage: boolean, path: string[], limitItem: number, cb: (dt: any, endReach: boolean) => void, err?: (error: any) => void) {
+    Firestore.paginate(isStartPage, path, [], [], limitItem, cb, err)
+  },
 }
 
 export default Firestore
